requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
Authlib==1.3.2
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Django==5.0.2
django-environ==0.11.2
django-htmx==1.17.2
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
dparse==0.6.4b0
filelock==3.12.4
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
packaging==23.2
phonenumbers==8.13.30
pillow==10.2.0
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
razorpay==1.4.1
requests==2.31.0
rich==13.8.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.2.7
safety-schemas==0.0.5
shellingham==1.5.4
sqlparse==0.4.4
stripe==8.4.0
tomli==2.0.1
typer==0.12.5
typing_extensions==4.9.0
urllib3==2.2.1
whitenoise==6.6.0
